[refactor] entity 추가
