+++ app/src/main/java/com/sopt/now/data/dto/response/ResponseFriendsDto.kt
@@ -1,5 +1,6 @@
 package com.sopt.now.data.dto.response
 
+import com.sopt.now.domain.entity.FriendEntity
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -14,24 +15,10 @@ data class ResponseFriendsDto(
     @SerialName("total_pages")
     val totalPages: Int,
     @SerialName("data")
-    val `data`: List<Data>,
+    val `data`: List<FriendEntity>,
     @SerialName("support")
     val support: Support
 ) {
-    @Serializable
-    data class Data(
-        @SerialName("id")
-        val id: Int,
-        @SerialName("email")
-        val email: String,
-        @SerialName("first_name")
-        val firstName: String,
-        @SerialName("last_name")
-        val lastName: String,
-        @SerialName("avatar")
-        val avatar: String
-    )
-
     @Serializable
     data class Support(
         @SerialName("url")
