+++ app/src/main/java/com/sopt/now/domain/entity/FriendEntity.kt
@@ -1,0 +1,18 @@
+package com.sopt.now.domain.entity
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class FriendEntity(
+    @SerialName("id")
+    val id: Int,
+    @SerialName("email")
+    val email: String,
+    @SerialName("first_name")
+    val firstName: String,
+    @SerialName("last_name")
+    val lastName: String,
+    @SerialName("avatar")
+    val avatar: String
+)
